Pipe: dynamically adjust connection timeout to handle SocketTimeoutException & Avoid resource cleaning when load task is in process (#12485)
* Pipe: dynamically adjust connection timeout to handle SocketTimeoutException

* Load: dynamically adjust connection timeout to handle SocketTimeoutException & Avoid resource cleaning when load task is in process

+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/receiver/protocol/airgap/IoTDBAirGapReceiver.java
@@ -68,7 +68,7 @@ public class IoTDBAirGapReceiver extends WrappedRunnable {
 
   @Override
   public void runMayThrow() throws Throwable {
-    socket.setSoTimeout((int) PipeConfig.getInstance().getPipeConnectorTransferTimeoutMs());
+    socket.setSoTimeout(PipeConfig.getInstance().getPipeConnectorTransferTimeoutMs());
     socket.setKeepAlive(true);
 
     LOGGER.info("Pipe air gap receiver {} started. Socket: {}", receiverId, socket);
